+++ src/main/java/org/cicirello/search/operators/permutations/CycleAlphaMutation.java
@@ -23,9 +23,10 @@ package org.cicirello.search.operators.permutations;
 import org.cicirello.search.operators.UndoableMutationOperator;
 import org.cicirello.permutations.Permutation;
 import org.cicirello.math.rand.RandomIndexer;
+import java.util.concurrent.ThreadLocalRandom;
 
 /**
- * <p>This class implements the Cycle(kmax) form of cycle mutation on permutations, where one mutation
+ * <p>This class implements the Cycle(&alpha;) form of cycle mutation on permutations, where one mutation
  * generates a random permutation cycle. Given the original parent permutation and
  * its mutant, a permutation cycle can be defined as follows. Imagine a graph with
  * n vertexes, where n is the permutation length. Now consider that for each index
@@ -38,23 +39,25 @@ import org.cicirello.math.rand.RandomIndexer;
  * a second example, p1 = [0, 1, 2, 3, 4], and p2 = [0, 4, 2, 1, 3]. This example has
  * a 3-cycle consisting of elements 1, 3, and 4.  Notice that position 1 has elements 1 and 4,
  * position 4 has elements 4 and 3, and position 3 has elements 3 and 1, so in the
- * hypothetical graph described above, there would be an edges from 1 to 4, 4 to 3, and 3 to 1,
+ * hypothetical graph described above, there would be edges from 1 to 4, 4 to 3, and 3 to 1,
  * a cycle of length 3.</p>
  *
- * <p>The Cycle(kmax) version of cycle mutation chooses the cycle size
- * k uniformly at random from [2, kmax], and then creates a random k-element cycle. The kmax is
- * a parameter allowing the evolutionary algorithm implementer the ability to control
- * the allowed range of cycle sizes. The combination of k elements is chosen uniformly 
- * at random from all possible combinations of k elements. Note that a 2-cycle is simply a swap.</p>
+ * <p>The Cycle(&alpha;) version of cycle mutation chooses the cycle size randomly from {2, 3, ..., n} where
+ * cycle length k is chosen with probability proportional to &alpha;<sup>k-2</sup>. It then generates
+ * a random permutation cycle of length k. The combination of k elements is chosen uniformly at 
+ * random from all possible combinations
+ * of k elements. Note that a 2-cycle is simply a swap.</p>
  *
- * <p>The runtime of
- * the {@link #mutate(Permutation) mutate} method is O(min(n, kmax<sup>2</sup>)), and derives
- * from the combination of algorithms utilized by the {@link RandomIndexer RandomIndexer}
- * class in sampling k random integers. For small values of kmax, the runtime is essentially
- * constant. The runtime of the {@link #undo(Permutation) undo} method is
- * O(kmax).</p>
+ * <p>The worst case runtime of a single call to 
+ * the {@link #mutate(Permutation) mutate} method is O(n), which occurs when the randomly
+ * chosen cycle length is n. However, this is a very low probability event. Lower cycle lengths
+ * are given significantly higher probability.
+ * The average case runtime of a single call to the 
+ * {@link #mutate(Permutation) mutate} method is O(min(n, ((2-&alpha;)/(1-&alpha;))<sup>2</sup>)).
+ * Thus, provided &alpha; is not close to 1, the average runtime is a constant depending upon 
+ * the value of &alpha;.</p>
  *
- * <p>Cycle mutation in both of its forms, including Cycle(kmax), was introduced in the following article:</p>
+ * <p>Cycle mutation in both of its forms, including Cycle(&alpha;), was introduced in the following article:</p>
  *
  * <p>Vincent A. Cicirello. 2022. <a href="https://www.cicirello.org/publications/applsci-12-05506.pdf">Cycle 
  * Mutation: Evolving Permutations via Cycle Induction</a>, <i>Applied Sciences</i>, 12(11), Article 5506 (June 2022). 
@@ -63,20 +66,25 @@ import org.cicirello.math.rand.RandomIndexer;
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
  */
-public final class CycleMutation implements UndoableMutationOperator<Permutation> {
+public final class CycleAlphaMutation implements UndoableMutationOperator<Permutation> {
 	
 	private int[] indexes;
-	private final int bound;
+	private final double logAlpha;
+	private final double alpha;
+	
+	// only recompute if permutation length different than last call to mutate
+	private int lastN;
+	private double term;
 	
 	/**
-	 * Constructs an CycleMutation mutation operator.
-	 * @param kmax The maximum length cycle to generate,
-	 * which must be at least 2.
-	 * @throws IllegalArgumentException if kmax &lt; 2.
+	 * Constructs an CycleAlphaMutation mutation operator.
+	 * @param alpha The alpha parameter of the mutation operator (see class documentation).
+	 * @throws IllegalArgumentException if alpha is less than or equal to 0 or greater than or equal to 1.
 	 */
-	public CycleMutation(int kmax) {
-		if (kmax < 2) throw new IllegalArgumentException("kmax too low");
-		bound = kmax - 1;
+	public CycleAlphaMutation(double alpha) {
+		if (alpha <= 0 || alpha >= 1) throw new IllegalArgumentException("alpha is outside allowed range");
+		logAlpha = Math.log(alpha);
+		this.alpha = alpha;
 	}
 	
 	@Override
@@ -84,7 +92,7 @@ public final class CycleMutation implements UndoableMutationOperator<Permutation
 		if (c.length() >= 2) {
 			indexes = RandomIndexer.sample(
 				c.length(),
-				2 + RandomIndexer.nextInt(bound < c.length() ? bound : c.length() - 1),
+				computeK(c.length(), ThreadLocalRandom.current().nextDouble()),
 				(int[])null
 			);
 			if (indexes.length > 2) {
@@ -118,8 +126,22 @@ public final class CycleMutation implements UndoableMutationOperator<Permutation
 	}
 	
 	@Override
-	public CycleMutation split() {
-		return new CycleMutation(bound + 1);
+	public CycleAlphaMutation split() {
+		return new CycleAlphaMutation(alpha);
+	}
+	
+	/*
+	 * package access to support testing
+	 */
+	final int computeK(int n, double u) {
+		if (n != lastN) {
+			term = 1 - Math.pow(alpha, n-1);
+			lastN = n;
+		}
+		int k = (int)(Math.log(1 - u * term) / logAlpha) + 2;
+		// things get numerically funny in the logs when alpha very near 1 and u also very near 1
+		// the check in the return protects against this
+		return k > n ? n : k;
 	}
 }