+++ CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2022-04-16
+## [Unreleased] - 2022-05-30
 
 ### Added
+* CycleAlphaMutation: an implementation of the Cycle(alpha) form of cycle mutation.
 
 ### Changed
 * Refactored hierarchy and methods of stochastic sampling classes.